+++ src/lib.rs
@@ -1,11 +1,11 @@
 //! Pythonic interface for Rust.
 
-mod enumerate;
 mod input;
+mod iterable;
 pub mod pathlib;
 pub mod string;
 pub mod time;
 
-pub use crate::enumerate::enumerate;
-pub use crate::enumerate::Iterable;
 pub use crate::input::input;
+pub use crate::iterable::enumerate;
+pub use crate::iterable::Iterable;
